+++ has_wall.c
@@ -6,7 +6,7 @@
 /*   By: ccardozo <ccardozo@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/09/04 12:23:32 by ccardozo          #+#    #+#             */
-/*   Updated: 2020/12/30 08:40:49 by ccardozo         ###   ########.fr       */
+/*   Updated: 2021/01/25 11:21:19 by ccardozo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,8 +17,8 @@ int		wall_colision(t_game *pos, float y, float x)
 	int		coly;
 	int		colx;
 
-	coly = (int)y / pos->tile.size;
-	colx = (int)x / pos->tile.size;
+	coly = (int)y / (int)pos->tile.size;
+	colx = (int)x / (int)pos->tile.size;
 	if (pos->map[coly][colx] == '1')
 		return (1);
 	return (0);
